Refactor common components of issue and epic detail view
and remove unused code
