+++ src/components/DetailView/Components/IssueSummary.tsx
@@ -3,32 +3,34 @@ import { Issue } from "types";
 import { useState } from "react";
 import { showNotification } from "@mantine/notifications";
 import { useMutation } from "@tanstack/react-query";
-import { editIssue } from "../helpers/queryFunctions";
 
 export function IssueSummary({
   summary,
   issueKey,
+  onMutate = () => {},
 }: {
   summary: string,
   issueKey: string,
+  onMutate?: () => void,
 }) {
-  const [defaultSummary, setDefaultSummary] = useState(summary);
+  const [currentSummary, setCurrentSummary] = useState(summary);
   const [showSummaryInput, setShowSummaryInput] = useState(false);
 
   const mutationSummary = useMutation({
-    mutationFn: (issue: Issue) => editIssue(issue, issueKey),
+    mutationFn: (issue: Issue) => window.provider.editIssue(issue, issueKey),
     onError: () => {
       showNotification({
-        message: "An error occurred while modifing the summary 😢",
+        message: "An error occurred while modifying the summary 😢",
         color: "red",
       });
     },
     onSuccess: () => {
-      if (defaultSummary !== summary) {
+      if (currentSummary !== summary) {
         showNotification({
           message: `The summary of issue ${issueKey} has been modified!`,
           color: "green",
         });
+        onMutate();
       }
     },
   });
@@ -36,26 +38,24 @@ export function IssueSummary({
   if (!showSummaryInput) {
     return (
       <Text lineClamp={1} onClick={() => setShowSummaryInput(true)}>
-        {defaultSummary}
+        {currentSummary}
       </Text>
     );
   }
 
   return (
     <Textarea
-      value={defaultSummary}
-      onChange={(e) => setDefaultSummary(e.target.value)}
+      value={currentSummary}
+      onChange={(e) => setCurrentSummary(e.target.value)}
       onBlur={() => {
-        if (defaultSummary === "") {
+        if (currentSummary === "") {
           showNotification({
             message: "The summary of an issue cannot be empty",
             color: "red",
           });
         } else {
           setShowSummaryInput(false);
-          mutationSummary.mutate({
-            summary: defaultSummary,
-          } as Issue);
+          mutationSummary.mutate({ summary: currentSummary } as Issue);
         }
       }}
       autosize
