+++ src/components/DetailView/Components/StoryPointsEstimateMenu.tsx
@@ -4,7 +4,6 @@ import { showNotification } from "@mantine/notifications";
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { Issue } from "types";
 import { getEditableIssueFields } from "../../CreateIssue/queryFunctions";
-import { editIssue } from "../helpers/queryFunctions";
 
 export function StoryPointsEstimateMenu({
   issueKey,
@@ -33,7 +32,7 @@ export function StoryPointsEstimateMenu({
   });
 
   const mutation = useMutation({
-    mutationFn: (issue: Issue) => editIssue(issue, issueKey),
+    mutationFn: (issue: Issue) => window.provider.editIssue(issue, issueKey),
     onError: () => {
       showNotification({
         message: `The story point estimate for issue ${issueKey} couldn't be modified! 😢`,
