+++ src/components/EpicDetailView/Components/ChildIssues.tsx
@@ -1,23 +1,17 @@
 import { Box, Button, Divider, ScrollArea, Stack } from "@mantine/core";
 import { useState } from "react";
-import { Issue } from "../../../../../types";
-import { CreateIssueModal } from "../../../CreateIssue/CreateIssueModal";
-import { ChildIssueWrapper } from "./ChildIssueWrapper";
-import { useColorScheme } from "../../../../common/color-scheme";
+import { Issue } from "../../../../types";
+import { CreateIssueModal } from "../../CreateIssue/CreateIssueModal";
+import { useColorScheme } from "../../../common/color-scheme";
+import { ChildIssueCard } from "./ChildIssueCard";
 
 export function ChildIssues({ issues }: { issues: Issue[] }) {
   const colorScheme = useColorScheme();
   const [createIssueModalOpened, setCreateIssueModalOpened] = useState(false);
 
   return (
-    <Stack
-      style={{
-        minHeight: "100%",
-        alignItems: "center",
-      }}
-    >
+    <Stack style={{ minHeight: "100%", alignItems: "center" }}>
       <ScrollArea
-        className="left-panel"
         p="sm"
         style={{
           minWidth: "260px",
@@ -26,7 +20,11 @@ export function ChildIssues({ issues }: { issues: Issue[] }) {
         }}
       >
         <Box mr="md">
-          <ChildIssueWrapper issues={issues} />
+          <Stack gap="sm">
+            {issues.map((issue: Issue, index) => (
+              <ChildIssueCard {...issue} key={issue.issueKey} index={index} />
+            ))}
+          </Stack>
         </Box>
         <Box mr="xs">
           <Button
@@ -56,13 +54,7 @@ export function ChildIssues({ issues }: { issues: Issue[] }) {
           setOpened={setCreateIssueModalOpened}
         />
       </ScrollArea>
-      <Divider
-        mr="xs"
-        size="xl"
-        style={{
-          cursor: "col-resize",
-        }}
-      />
+      <Divider mr="xs" size="xl" style={{ cursor: "col-resize" }} />
     </Stack>
   );
 }
