--- src/components/EpicDetailView/helpers/queryFunctions.ts
@@ -1,3 +1,0 @@
-import { Issue } from "types";
-
-export const editIssue = (issue: Issue, issueIdOrKey: string): Promise<void> => window.provider.editIssue(issue, issueIdOrKey);
